move database helper
