+++ src/main/java/edu/byu/cs/autograder/database/DatabaseHelper.java
@@ -1,5 +1,7 @@
-package edu.byu.cs.autograder;
+package edu.byu.cs.autograder.database;
 
+import edu.byu.cs.autograder.GradingContext;
+import edu.byu.cs.autograder.GradingException;
 import edu.byu.cs.properties.ApplicationProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
